+++ store/cachemulti/store.go
@@ -50,16 +50,15 @@ func NewFromKVStore(
 
 	for key, store := range stores {
 		var cacheWrapped types.CacheWrap
-		if cms.TracingEnabled() {
+		switch {
+		case cms.ListeningEnabled(key):
+			cacheWrapped = store.CacheWrapWithListeners(key, cms.listeners[key])
+		case cms.TracingEnabled():
 			cacheWrapped = store.CacheWrapWithTrace(cms.traceWriter, cms.traceContext)
-		} else {
+		default:
 			cacheWrapped = store.CacheWrap()
 		}
-		if cms.ListeningEnabled(key) {
-			cms.stores[key] = cacheWrapped.CacheWrapWithListeners(key, cms.listeners[key])
-		} else {
-			cms.stores[key] = cacheWrapped
-		}
+		cms.stores[key] = cacheWrapped
 	}
 
 	return cms